backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
sqlalchemy==2.0.23
pydantic==2.4.2
python-multipart==0.0.6
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
groq==0.4.1
python-dotenv==1.0.0
pillow==10.0.1
aiofiles==23.2.1